requirements.txt
# Shamela Scraper v2.0 - Required Libraries
# المكتبات المطلوبة لمشروع Shamela Scraper

# Web Scraping & HTTP
aiohttp>=3.8.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Database
mysql-connector-python>=8.0.0

# Data Analysis & Visualization
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0

# System & Performance
psutil>=5.9.0

# Network & URL handling
urllib3>=1.26.0

# Additional utilities
argparse  # Built-in but listed for completeness

# Development & Testing (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0

# Note: The following are built-in Python modules and don't need installation:
# - asyncio
# - json
# - time
# - os
# - sys
# - re
# - logging
# - threading
# - queue
# - subprocess
# - hashlib
# - sqlite3
# - tkinter
# - pathlib
# - datetime
# - typing
# - dataclasses
# - concurrent.futures
# - multiprocessing
# - functools
# - io
# - gzip
# - pickle
# - contextlib
# - traceback
# - statistics
# - shutil